refactor: create new `app.js` for alpine; save old code as backup in `app.legacy.js`
